==> bert/4_inference.ipynb <==
import transformers

sentence = 'FOB DJIBOUTI / ETHIOPIA PORT QTY : 363 MT ( 15 X 20 FT CONTAINERS EACH IS STUFFED WITH 24MTS ) BOLD DESI CHICK PEAS, NEW CROP 2018 / 19, PURITY : 98.5 PERCENT SHAWA TYPE, MACHINES CLEANED, TOTAL DEFECTS MAX 1.5 PCT AND MAX 2 PCT BLACK, PACKED IN 50KG PP WOOVEN BAGS, AT USD 555 PER M / TON. AS PER BENEFICIARY SALES CONTRACT NO. 1 FNTTP / 007 / 2019 DATED 26.11.19'

# LOAD TOKENIZER
from transformers import PreTrainedTokenizerFast

tokenizer = PreTrainedTokenizerFast(
    tokenizer_file="tokenizer/tokenizer_v2.json",
    bos_token="[S]",
    eos_token="[/S]",
    unk_token="[UNK]",
    pad_token="[PAD]",
    cls_token="[CLS]",
    sep_token="[SEP]",
    mask_token="[MASK]",
    padding_side="right",
    model_max_len=300
)

from transformers import TFBertForTokenClassification

model = TFBertForTokenClassification.from_pretrained('./bert-ner-conpend-300/')

from transformers import pipeline

nlp = pipeline('ner', model=model, tokenizer=tokenizer)
result = nlp(sentence)
result

print(result)
